# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1.0.0",
    "title": "Microservice Posts",
    "description": "Microservice from the \"If community\" project for managing posts and their categories and interactions (upvotes)"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "api-token",
      "description": "An API token is a unique access key that grants permission to interact with the API, ensuring security and control over the operations performed."
    }
  },
  "paths": {
...
    "/categories": {
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "adding new category to microservice",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "category data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tecnologia da informação"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tecnologia da informação"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "id": {
                  "type": "number",
                  "example": 3
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-18T12:23:36.854Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Uma categoria com esse nome já foi cadastrada no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Conflict"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "lists all categories registered in the system",
        "parameters": [
          {
            "name": "pageNumber",
            "in": "query",
            "description": "Number informing which page the search should return\n (pattern 1)",
            "type": "number"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Number informing the number of elements per page\n (pattern 10)",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "totalPages": {
                  "type": "number",
                  "example": 1
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 1
                      },
                      "createdAt": {
                        "type": "string",
                        "example": "2024-05-19T13:38:28.030Z"
                      },
                      "updatedAt": {
                        "type": "string",
                        "example": "null"
                      },
                      "deletedAt": {
                        "type": "string",
                        "example": "null"
                      },
                      "name": {
                        "type": "string",
                        "example": "Tecnologia"
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "search for a category by ID",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The category ID"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tecnologia da informação"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "id": {
                  "type": "number",
                  "example": 3
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-18T12:23:36.854Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Categoria com esse id não está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Categories"
        ],
        "description": "update category specified by ID",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The category ID"
          },
          {
            "name": "body",
            "in": "body",
            "description": "category data.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tecnologia da informação"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Tecnologia da informação"
                },
                "updatedAt": {
                  "type": "string",
                  "example": "null"
                },
                "id": {
                  "type": "number",
                  "example": 3
                },
                "createdAt": {
                  "type": "string",
                  "example": "2024-05-18T12:23:36.854Z"
                },
                "deletedAt": {
                  "type": "string",
                  "example": "null"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created"
          },
          "400": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "(error message) no (parameter)"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Categoria com esse id não está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Uma categoria com esse nome já foi cadastrada no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Conflict"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "removes a category from the specified id",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "number",
            "description": "The category ID"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "delete": {
                  "type": "boolean",
                  "example": true
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "OK"
          },
          "401": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Token inválido. Forneça um token de autenticação válido."
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Categoria com esse id não está cadastrado no sistema"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Not Found"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
...
  }
}